Minor: Fix #15 - When there's a new download, button opens a popup
This is a bug in vanilla AiOS. When there is a new download (i.e, saved something), downloads-button (the one overlayed by the ext.) gets hidden/collapsed and gets replaced by downloads-indicator whose oncommand launches a popup instead of opening it inside the sidebar/downloads window.

Solution: use a mutation observer to monitor if the downloads indicator element is now present since it doesn't appear unless a new download gets launched. to others, feel free to present a better solution to this.